Moved 2 manage_action_container tests to test_tags.py
